Handle link and image references
This adds handlers for the ImageReference and LinkReference nodes to
resolve them down to Slack-compatible links. It doesn't seem possible to
look up Definition nodes from the handler arguments, so we crawl the AST
beforehand to manually build up our own record, then remove the nodes so
that they do not clog up the output with awkward newlines.

Relevant nodes:

- https://github.com/syntax-tree/mdast/blob/main/readme.md#definition
- https://github.com/syntax-tree/mdast/blob/main/readme.md#linkreference
- https://github.com/syntax-tree/mdast/blob/main/readme.md#imagereference